Add stub implementation and test for bulk mail scrubbing
