+++ tests/lib/scrubbing/test_mail_scrubbing.py
@@ -1,16 +1,10 @@
-# TODO archivable member
-from datetime import datetime
-
 import pytest
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
-from pycroft.helpers.interval import closedopen
-from pycroft.lib.user_deletion import scrub_mail, scrubbable_mails
-from pycroft.model.config import Config
+from pycroft.lib.user_deletion import scrub_all_mails, scrub_mail, scrubbable_mails
 from pycroft.model.scrubbing import ScrubLog
-from pycroft.model.user import PropertyGroup, User
-from tests import factories as f
+from pycroft.model.user import User
 from tests.assertions import assert_one
 
 
@@ -40,35 +34,7 @@ def test_scrubbing_scrubs_mail(user_archivable: User, processor: User, session: 
             )
 
 
-@pytest.fixture(scope="module")
-def user_archivable(module_session: Session, config: Config) -> User:
-    # TODO: user wtih membership
-    return f.UserFactory.create(
-        registered_at=datetime(2020, 7, 1),
-        with_membership=True,
-        membership__active_during=closedopen(datetime(2020, 7, 1), datetime(2021, 11, 25)),
-        membership__group=config.member_group,
-        without_room=True,
-    )
-
-
-@pytest.fixture(scope="module", autouse=True)
-def user_do_not_archive(
-    module_session: Session, config: Config, group_do_not_archive: PropertyGroup
-) -> User:
-    """
-    Create a user with a membership in a group that has the do-not-archive
-    property.
-    """
-    # TODO _very old_ membership, but also mem in a group
-    user = f.UserFactory.create(
-        registered_at=datetime(2020, 7, 1),
-        with_membership=True,
-        membership__active_during=closedopen(datetime(2020, 7, 1), datetime(2021, 11, 25)),
-        membership__group=config.member_group,
-        without_room=True,
-    )
-    f.MembershipFactory.create(
-        user=user, group=group_do_not_archive, active_during=closedopen(datetime(2020, 7, 1), None)
-    )
-    return user
+def test_bulk_scrubbing_scrubs_mail(
+    users_archivable: list[User], users_notarchivable: list[User], processor: User, session: Session
+):
+    scrub_all_mails(session, author=processor)
